combine auth HOCs
