+++ config/globalEvents.json
@@ -1,0 +1,4 @@
+{
+  "SIGN_IN_EVENT": "SIGN_IN_EVENT",
+  "SIGN_OUT_EVENT": "SIGN_OUT_EVENT"
+}
