+++ lib/apollo/hooks/actions.js
@@ -1,6 +1,7 @@
 import { useMutation } from '@apollo/client';
 import { Router } from 'routes';
 import { HOME } from 'config/routes';
+import { SIGN_IN_EVENT, SIGN_OUT_EVENT } from 'config/globalEvents.json';
 
 import SignIn from 'graphql/mutations/signIn.graphql';
 import SignUp from 'graphql/mutations/signUp.graphql';
@@ -26,6 +27,8 @@ export const useSignIn = () => {
   const mutate = async ({ email, password }) => {
     try {
       await mutation({ variables: { email, password } });
+
+      window.localStorage.setItem(SIGN_IN_EVENT, Date.now());
 
       Router.pushRoute(HOME.pattern);
     } catch (error) {
@@ -54,6 +57,8 @@ export const useSignUp = () => {
     try {
       await mutation({ variables: { email, password, ...rest } });
 
+      window.localStorage.setItem(SIGN_IN_EVENT, Date.now());
+
       Router.pushRoute(HOME.pattern);
     } catch (error) {
       console.log(error);
@@ -78,6 +83,8 @@ export const useSignOut = () => {
   const mutate = async () => {
     try {
       await mutation();
+
+      window.localStorage.setItem(SIGN_OUT_EVENT, Date.now());
 
       Router.pushRoute(HOME.pattern);
     } catch (error) {
